refactor check_players_guess
When reading check_players_guess the english of your methods made sense:
set_correct_guesses, set_incorrect_guesses and
set_correct_letters_if_correct_guess.

But after reading I didn't know what it was doing without diggin into
each of those methods. Also it was doing two things. First it was
checking if the guess was right and classifying the letter as correct or
incorrect. Second it was building up the somewhat more complex
correct_letters array, which seemed like a different behavior.

Since correct_letters is only used in the display I moved that logic
into guessed_word which we can call when we need to re-draw the display.