rename random_word to correct_word
While true it was a random_word, in all the subsequent contexts it is
used in comparison against the guesses to see if the guess is correct.

I also privatized the accessors since they are not used outside the Game
class.

And removed create_correct_letters_array, it wasn't being used anymore.